transfer add book + Fix remove book function
